Remove calls to get thumbnail post.
When using this plugin with
https://github.com/humanmade/network-media-library it doesn't work, it
seems this is because it attempts to retrieve the post of the thumbnail
(needlessly afaict) which won't work because the image exists on a
different site in the network.

Removing these calls seems to fix the issue and allows for a better
mutli site experience. I'm not sure of any adverse affects of this, but
let me know if you can see one.